fix: add error handling for loadStateFromComment and extract state encoder
- Wrap loadStateFromComment() in try-catch to prevent CLI crash on
  GitHub API failures, falling back to fresh initial state
- Extract embedState()/extractState() into dedicated comment-state.ts
  module to separate state serialization from rendering concerns
